requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
SQLAlchemy==2.0.46
asyncpg==0.31.0
alembic==1.18.3
python-dotenv==1.2.1
pydantic==2.12.5